==> CDD_training_aids/New_Stuff/Tissue/1_Load_Review_for_tissue-copy.Rmd <==
---
title: "Load & Review Data v2 (get all Tissue sorted)"
author: "Wesley"
date: "2022/03/28"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Load all of the Tissue Data

To start, we need to load all of the Excel files. To do this, we will use the
readxl package. In addition, we're going to need to manipulate strings in 
the following, so let's pre-emptively load that package.

```{r libraries, error = FALSE, message = FALSE, warning = FALSE}
library("readxl")
library("stringr")
```

### Start with Tissue 1, as Template

Tissue 1 has many sheets, each of which is a nicely organized listing of VOCs
as detected by the GCxGC system's built-in scanning software, and exported. 
We will load them all. 

```{r load_data}
sheets <- excel_sheets("/home/wburr/Rushali_Analysis/New_Stuff/Tissue/Tissue1.xlsx")  # note: spaces in names
Tissue1 <- lapply(sheets, FUN = function(x) {
           read_xlsx(path = "/home/wburr/Rushali_Analysis/New_Stuff/Tissue/Tissue1.xlsx",
                     sheet = x) })
sheets <- str_replace_all(sheets, " ", "_")
names(Tissue1) <- sheets
```

Having things organized by sample is a good thing, but we really will
find it annoying finding things long-term. In the following, I will take each
of the lists of compounds relevant to Tissue 1, and merge them into a single
data.frame object, so we can then filter that frame down to get a specific
compound, or a specific day/observation, and so on.

```{r merge_sheets}
Tissue1m <- do.call("rbind", Tissue1)
names(Tissue1m) <- c("PeakNum", "Sample", "Name", "Formula", "RTs", "1stRT","2ndRT","Similarity",
                   "Area", "Height", "QuantMass", "BaseMass", "QuantSN", "PeakSN",
                   "Type", "Conc", "ConcUnits")
```

A quick print of the data structure shows that the RTs, QuantMass and BaseMass
won't be usable as numbers. The RTs need to be split, as well.

```{r split_RTs}
RT1 <- as.numeric(
         unlist(
           lapply(strsplit(Tissue1m$RTs, ","), FUN = "[[", 1)
         )
       )
RT2 <- as.numeric(
         unlist(
           lapply(strsplit(Tissue1m$RTs, ","), FUN = "[[", 2)
         )
       )
Tissue1m <- data.frame(Tissue1m[, c(1:4, 6:17)], RT1 = RT1, RT2 = RT2)
```

```{r samples}
unique(Tissue1m$Sample)
save(file = "/home/wburr/Rushali_Analysis/New_Stuff/Tissue/Tissue1.rda", Tissue1m)
```

### Now Do Tissue 2

```{r load_data2}
sheets <- excel_sheets("/home/wburr/Rushali_Analysis/New_Stuff/Tissue/Tissue2.xlsx")  # note: spaces in names
Tissue2 <- lapply(sheets, FUN = function(x) {
           read_xlsx(path = "/home/wburr/Rushali_Analysis/New_Stuff/Tissue/Tissue2.xlsx",
                     sheet = x) })
sheets <- str_replace_all(sheets, " ", "_")
names(Tissue2) <- sheets

Tissue2m <- do.call("rbind", Tissue2)
names(Tissue2m) <- c("PeakNum", "Sample", "Name", "Formula", "RTs","1stRT","2ndRT", "Similarity",
                   "Area", "Height", "QuantMass", "BaseMass", "QuantSN", "PeakSN",
                   "Type", "Conc", "ConcUnits")
```

```{r split_RTs2}
RT1 <- as.numeric(
         unlist(
           lapply(strsplit(Tissue2m$RTs, ","), FUN = "[[", 1)
         )
       )
RT2 <- as.numeric(
         unlist(
           lapply(strsplit(Tissue2m$RTs, ","), FUN = "[[", 2)
         )
       )
Tissue2m <- data.frame(Tissue2m[, c(1:4, 6:17)], RT1 = RT1, RT2 = RT2)
```

```{r samples2}
unique(Tissue2m$Sample)
save(file = "/home/wburr/Rushali_Analysis/New_Stuff/Tissue/Tissue2.rda", Tissue2m)
```

### Now Do Tissue 3

```{r load_data3}
sheets <- excel_sheets("/home/wburr/Rushali_Analysis/New_Stuff/Tissue/Tissue3.xlsx")  # note: spaces in names
Tissue3 <- lapply(sheets, FUN = function(x) {
           read_xlsx(path = "/home/wburr/Rushali_Analysis/New_Stuff/Tissue/Tissue3.xlsx",
                     sheet = x) })
sheets <- str_replace_all(sheets, " ", "_")
names(Tissue3) <- sheets

Tissue3m <- do.call("rbind", Tissue3)
names(Tissue3m) <- c("PeakNum", "Sample", "Name", "Formula", "RTs","1stRT","2ndRT", "Similarity",
                   "Area", "Height", "QuantMass", "BaseMass", "QuantSN", "PeakSN",
                   "Type", "Conc", "ConcUnits")
```

```{r split_RTs3}
RT1 <- as.numeric(
         unlist(
           lapply(strsplit(Tissue3m$RTs, ","), FUN = "[[", 1)
         )
       )
RT2 <- as.numeric(
         unlist(
           lapply(strsplit(Tissue3m$RTs, ","), FUN = "[[", 2)
         )
       )
Tissue3m <- data.frame(Tissue3m[, c(1:4, 6:17)], RT1 = RT1, RT2 = RT2)
```

```{r samples3}
unique(Tissue3m$Sample)
save(file = "/home/wburr/Rushali_Analysis/New_Stuff/Tissue/Tissue3.rda", Tissue3m)
```
